Fix a bug where assemble would die if it tried to process any unreadable PDFs.
It wass because the gs feeding process knew that we skipped the corrupt file
but the gs consuming process had no way to find that out.  All further
bookmarks would be wrong, then we'd die with a "file too long" error.
Fixed  by getting ghostscript to always print the name of the file that
it's starting to process.